requirements.txt
Flask
pyodbc
python-dotenv
gunicorn